pyproject.toml
[project]
name = "agentlens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "ddgs>=9.12.1",
    "ollama>=0.6.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "pandas>=2.0.0",
    "pandas>=2.0.0",
]
